change interface to client, make test impl local
